+++ pkg/kcp/iprange/state.go
@@ -3,7 +3,7 @@ package iprange
 import (
 	cloudresourcesv1beta1 "github.com/kyma-project/cloud-manager/components/kcp/api/cloud-control/v1beta1"
 	"github.com/kyma-project/cloud-manager/components/kcp/pkg/common/actions/focal"
-	"github.com/kyma-project/cloud-manager/components/kcp/pkg/iprange/types"
+	"github.com/kyma-project/cloud-manager/components/kcp/pkg/kcp/iprange/types"
 )
 
 type state struct {
