+++ pkg/provider/gcp/iprange/new.go
@@ -3,11 +3,11 @@ package iprange
 import (
 	"context"
 	"fmt"
+	"github.com/kyma-project/cloud-manager/components/kcp/pkg/kcp/iprange/types"
 
 	"github.com/kyma-project/cloud-manager/components/kcp/api/cloud-control/v1beta1"
 	"github.com/kyma-project/cloud-manager/components/kcp/pkg/common/actions"
 	"github.com/kyma-project/cloud-manager/components/kcp/pkg/composed"
-	"github.com/kyma-project/cloud-manager/components/kcp/pkg/iprange/types"
 	"github.com/kyma-project/cloud-manager/components/kcp/pkg/provider/gcp/client"
 )
